requirements.txt
djangorestframework
django-filter
django-cors-headers
markdown